factor out helper objects from sqlite3_inspector's file
