+++ tests/skibidi-orm/migration_engine/db_inspector/test_db_inspectors.py
@@ -4,9 +4,10 @@ from skibidi_orm.migration_engine.db_config.sqlite3_config import SQLite3Config
 from skibidi_orm.migration_engine.adapters.database_objects import constraints as c
 import sqlite3
 
-from skibidi_orm.migration_engine.db_inspectors.sqlite3_inspector import (
+from skibidi_orm.migration_engine.db_inspectors.sqlite.sqlite3_inspector import (
     SQLite3Inspector,
 )
+
 
 sql_table1 = """
     CREATE TABLE table1 (
